feat: introduce textual refiner workspace
