+++ Arcanalte.py
@@ -17,8 +17,7 @@ from arcana.pages.chatbot import chatbot_page
 from arcana.pages.settings import settings_page
 from arcana.pages.mixup import mixup_page
 from arcana.pages.longresponse import longresponse_page
-from arcana.pages.editor import editor_page
-from arcana.pages.refiner import grammar_refiner_page
+from arcana.pages.textual_refiner import textual_refiner_page
 
 # Import configurations
 from arcana.core.config import APP_TITLE, CACHE_DIR, INDEX_FILE
@@ -86,7 +85,7 @@ st.sidebar.title("Navigation")
 # --- Page Selection Logic ---
 # Define the pages
 main_pages = ["Introduction", "Files", "Citations", "Chatbot"]
-advanced_pages = ["Mixup", "Editor", "Long Response", "Grammar Refiner"]
+advanced_pages = ["Mixup", "Textual Refiner", "Long Response"]
 
 # Display main page buttons
 for page in main_pages:
@@ -151,12 +150,11 @@ def citations_page():
 pages = {
     "Introduction": intro_page,
     "Files": files_page,
-    "Citations": citations_page, 
+    "Citations": citations_page,
     "Chatbot": chatbot_page,
     "Mixup": mixup_page,
-    "Editor": editor_page,
+    "Textual Refiner": textual_refiner_page,
     "Long Response": longresponse_page,
-    "Grammar Refiner": grammar_refiner_page,
     "Settings": settings_page
 }
