+++ arcana/pages/__init__.py
@@ -1,6 +1,5 @@
 """UI page components for the Arcana Streamlit app."""
 
-from .editor import editor_page
-from .refiner import grammar_refiner_page
+from .textual_refiner import textual_refiner_page
 
-__all__ = ["editor_page", "grammar_refiner_page"]
+__all__ = ["textual_refiner_page"]
